Use Int32 for line numbers

+++ swift/src/Ice/Incoming.swift
@@ -222,32 +222,31 @@ public final class Incoming {
     }
   }
 
-  // TODO: the line number in LocalException should be an Int32, not an Int
   func convertIntoDispatchException(_ exception: Error) -> ICEDispatchException {
     switch exception {
     // OperationNotExistException and friends
     case let e as ObjectNotExistException:
       ICEDispatchException.objectNotExistException(
         e.id.name, category: e.id.category, facet: e.facet, operation: e.operation, file: e.file,
-        line: Int32(e.line))
+        line: e.line)
     case let e as FacetNotExistException:
       ICEDispatchException.facetNotExistException(
         e.id.name, category: e.id.category, facet: e.facet, operation: e.operation, file: e.file,
-        line: Int32(e.line))
+        line: e.line)
     case let e as OperationNotExistException:
       ICEDispatchException.operationNotExistException(
         e.id.name, category: e.id.category, facet: e.facet, operation: e.operation, file: e.file,
-        line: Int32(e.line))
+        line: e.line)
     // Unknown exceptions
     case let e as UnknownUserException:
-      ICEDispatchException.unknownUserException(e.unknown, file: e.file, line: Int32(e.line))
+      ICEDispatchException.unknownUserException(e.unknown, file: e.file, line: e.line)
     case let e as UnknownLocalException:
-      ICEDispatchException.unknownLocalException(e.unknown, file: e.file, line: Int32(e.line))
+      ICEDispatchException.unknownLocalException(e.unknown, file: e.file, line: e.line)
     case let e as UnknownException:
-      ICEDispatchException.unknownException(e.unknown, file: e.file, line: Int32(e.line))
+      ICEDispatchException.unknownException(e.unknown, file: e.file, line: e.line)
     // Other exceptions mapped to Unknown exceptions
     case let e as LocalException:
-      ICEDispatchException.unknownLocalException("\(e)", file: e.file, line: Int32(e.line))
+      ICEDispatchException.unknownLocalException("\(e)", file: e.file, line: e.line)
     case let e as UserException:
       ICEDispatchException.unknownUserException("\(e.ice_id())", file: #file, line: #line)
     default:
